backend/testApp/pyproject.toml
[project]
name = "testapp"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.115.6,<0.116.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "black (>=24.10.0,<25.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
